+++ Tickets/src/main/java/com/tickets/Tickets/service/PlanService.java
@@ -13,7 +13,7 @@ public interface PlanService {
 	 * @date 2018年4月17日
 	 * 
 	 */
-	public List<Plan> getPlans(int pageSize, int index, Page page, String keyword, String day1, String day2,
+	public List<Plan> getPlans(Long pageSize, Long index, Page page, String keyword, String day1, String day2,
 			String location, String overdue, String isrecommend, String type, String sort_strategy);
 
 	/**
@@ -21,7 +21,7 @@ public interface PlanService {
 	 * @date 2018年4月21日
 	 * 
 	 */
-	public Plan getById(int planid);
+	public Plan getById(Long planid);
 	
 	
 }
